Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -1,6 +1,6 @@
 #include<stdio.h>
 int main(){
-    int n,m;
+    float n,m;
     scanf("%f %f", &n, &m);
     printf("Product: %.2f", n*m);
     return 0;
